py-backend/requirements.txt
APScheduler==3.9.1
Flask==2.2.2
Js2Py==0.74
PyYAML==6.0
requests==2.25.0
